Fix audio on rpi

+++ src/phazor/phazor.c
@@ -594,12 +594,12 @@ void read_to_buffer_char16_resample(char src[], int n_bytes) {
 
     // Convert bytes16 to float
     while (i < n_bytes) {
-        re_in[f * 2] = ( ((src[i + 1] << 8) | (src[i + 0] & 0xFF))) / 32768.0;
+        re_in[f * 2] = (float)((int16_t)((src[i + 1] << 8) | (src[i + 0] & 0xFF)) / 32768.0);
         if (src_channels == 1) {
             re_in[(f * 2) + 1] = re_in[f * 2];
             i += 2;
         } else {
-            re_in[(f * 2) + 1] = (((src[i + 3] << 8) | (src[i + 2] & 0xFF))) / 32768.0;
+            re_in[(f * 2) + 1] = (float)((int16_t)((src[i + 3] << 8) | (src[i + 2] & 0xFF)) / 32768.0);
             i += 4;
         }
 
@@ -622,7 +622,7 @@ void read_to_buffer_char16(char src[], int n_bytes) {
     int i = 0;
     if (src_channels == 1){
         while (i < n_bytes) {
-            bfl[high] = ((src[i + 1] << 8) | (src[i + 0] & 0xFF)) / 32768.0;
+            bfl[high] = (float)((int16_t)((src[i + 1] << 8) | (src[i + 0] & 0xFF)) / 32768.0);
             bfr[high] = bfl[high];
             if (fade_fill > 0 || fade_mini < 1.0) {
                 fade_fx();
@@ -632,8 +632,8 @@ void read_to_buffer_char16(char src[], int n_bytes) {
         }
     } else {
         while (i < n_bytes) {
-            bfl[high] = (((src[i + 1] << 8) | (src[i + 0] & 0xFF)) / 32768.0);
-            bfr[high] = (((src[i + 3] << 8) | (src[i + 2] & 0xFF)) / 32768.0);
+            bfl[high] = (float)((int16_t)((src[i + 1] << 8) | (src[i + 0] & 0xFF)) / 32768.0);
+            bfr[high] = (float)((int16_t)((src[i + 3] << 8) | (src[i + 2] & 0xFF)) / 32768.0);
             if (fade_fill > 0 || fade_mini < 1.0) {
                 fade_fx();
             }
